chore: split react docs logic into two files
